// app/requests/_personal/getSpecialistForBid/getSpecialistForBidSuccess.type.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface GetSpecialistForBidSuccess {
  data: {
    id: number;
    phone: number;
    email: string;
    logo?: string | null;
    roles: {
      id: number;
      name:
        | "admin"
        | "client"
        | "manager"
        | "recruiter"
        | "specialist"
        | "supervisor";
      [k: string]: unknown;
    }[];
    radius: string;
    name: string;
    age: string;
    country: string;
    viewActivities: string[];
    [k: string]: unknown;
  }[];
  [k: string]: unknown;
}